mon/mon_types: move C_MonOp to MonOpRequest.h
Fixes a circular #include of
 mon_types -> MonOpRequest -> Session -> mon_types
